@@ -13,10 +13,10 @@
     "apollo-boost": "^0.4.9",
     "apollo-link-token-refresh": "^0.7.0",
     "cloudinary-react": "^1.7.0",
+    "email-validator": "^2.0.4",
     "formik": "^2.2.6",
     "framer-motion": "^4",
     "graphql": "^15.5.0",
-    "jwt-decode": "^4.0.0",
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
     "radium": "^0.26.1",
@@ -51,6 +51,7 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@eslint/js": "^9.6.0",
     "@graphql-codegen/cli": "1.21.3",
     "@graphql-codegen/typescript": "1.21.1",
@@ -61,12 +62,11 @@
     "@testing-library/user-event": "^12.1.10",
     "@types/graphql": "^14.5.0",
     "@types/jest": "^26.0.15",
-    "@types/jwt-decode": "^3.1.0",
     "@types/node": "^20.14.10",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
     "@types/react-router-dom": "^5.1.7",
-    "eslint": "8.x",
+    "eslint": "^9.7.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.34.3",
